worker/requirements.txt
psycopg2-binary==2.9.9
vaderSentiment==3.3.2
